+++ c_sharp_for_home_assistant/src/hhnl.HomeAssistantNet/hhnl.HomeAssistantNet.Automations/Automation/Runner/AllowAutomationRunner.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Concurrent;
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using hhnl.HomeAssistantNet.Shared.Automation;
@@ -17,9 +18,10 @@ namespace hhnl.HomeAssistantNet.Automations.Automation.Runner
         public override Task EnqueueAsync(
             AutomationRunInfo.StartReason reason,
             string? changedEntity,
-            TaskCompletionSource? startTcs)
+            TaskCompletionSource? startTcs,
+            IReadOnlyDictionary<Type, object> snapshot)
         {
-            var run = CreateAutomationRun(reason, changedEntity, startTcs);
+            var run = CreateAutomationRun(reason, changedEntity, startTcs, snapshot);
             
             Entry.AddRun(run);
             run.Start();