+++ c_sharp_for_home_assistant/CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [Snapshot] attribute
+- ReentryPolicy.QueueLatest - Behaves similar to what ReentryPolicy.Queue was before
+
+### Changed
+- ReentryPolicy.Queue - Will now queue runs without a limit
 
 ## [0.4.2] - 2021-07-02
 ### Fixed